SM: auto_str length to int32_t
 - UTF8 helper uses negative return values
 - A string variable with int64 length is not realistic for the rest of
the framework

+++ src/jx/Proxy/Mozilla/MozJS/MozValue.h
@@ -14,7 +14,7 @@ class String;
   class name {                                                         \
    public:                                                             \
     type *str_;                                                        \
-    size_t length_;                                                    \
+    int32_t length_;                                                   \
     JSContext *ctx_;                                                   \
                                                                        \
     name() {                                                           \
